+++ packages/opencode/src/kilocode/index.ts
@@ -3,3 +3,4 @@ export { RulesMigrator } from "./rules-migrator"
 export { WorkflowsMigrator } from "./workflows-migrator"
 export { McpMigrator } from "./mcp-migrator"
 export { KilocodeConfigInjector } from "./config-injector"
+export { KilocodePaths } from "./paths"
